pyproject.toml
[tool.poetry]
name = "oi2cms"
version = "0.1.2"
description = ""
authors = ["Alireza Kaviani <alireza.kaviani.2005@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
oi2cms = "oi2cms.main:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.9.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
